fix: specify nuxt compatibility
