+++ src/types.d.ts
@@ -1,0 +1,9 @@
+export interface CustomMachinesOptions {
+    dir?: string,
+    importSuffix?: string,
+  }
+
+export interface ModuleOptions {
+    minimal: boolean;
+    customMachines: CustomMachinesOptions | false;
+  }
